bech32 address support in bitcore

+++ src/key_io.h
@@ -28,7 +28,7 @@ bool IsValidDestinationString(const std::string& str, const CChainParams& params
 bool IsValidContractSenderAddressString(const std::string& str);
 
 #ifdef ENABLE_BITCORE_RPC
-bool DecodeIndexKey(const std::string& str, uint160& hashBytes, int& type);
+bool DecodeIndexKey(const std::string& str, uint256& hashBytes, int& type);
 #endif
 
 #endif // BITCOIN_KEY_IO_H
